p1/requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pillow==10.0.0
matplotlib==3.7.2
tqdm==4.66.1
opencv-python==4.8.0.76
tensorboard==2.13.0
